refactor SetEncoder
