+++ knora/dsplib/utils/set_encoder.py
@@ -1,0 +1,8 @@
+import json
+
+
+class SetEncoder(json.JSONEncoder):
+    def default(self, obj):
+        if isinstance(obj, set):
+            return list(obj)
+        return json.JSONEncoder.default(self, obj)
